clock 알고리즘 및 기타 함수들  수정, 3개 fail

+++ include/vm/anon.h
@@ -7,7 +7,7 @@ enum vm_type;
 
 struct anon_page {
     // struct page anon_p; // heesan 주의☠️ ??
-    size_t swap_location;   // swap disk 위치
+    int swap_location;   // swap disk 위치
 };
 
 void vm_anon_init (void);
